+++ packages/protocol/contracts/L1/tiers/TierProviderV2.sol
@@ -4,10 +4,10 @@ pragma solidity 0.8.24;
 import "../../common/EssentialContract.sol";
 import "./ITierProvider.sol";
 
-/// @title TestnetTierProvider
+/// @title TierProviderV2
 /// @dev Labeled in AddressResolver as "tier_provider"
 /// @custom:security-contact security@taiko.xyz
-contract TestnetTierProvider is EssentialContract, ITierProvider {
+contract TierProviderV2 is EssentialContract, ITierProvider {
     uint256[50] private __gap;
 
     /// @notice Initializes the contract.
@@ -18,25 +18,25 @@ contract TestnetTierProvider is EssentialContract, ITierProvider {
 
     /// @inheritdoc ITierProvider
     function getTier(uint16 _tierId) public pure override returns (ITierProvider.Tier memory) {
-        if (_tierId == LibTiers.TIER_OPTIMISTIC) {
-            return ITierProvider.Tier({
-                verifierName: "tier_optimistic",
-                validityBond: 250 ether, // TKO
-                contestBond: 500 ether, // TKO
-                cooldownWindow: 1440, //24 hours
-                provingWindow: 30, // 0.5 hours
-                maxBlocksToVerifyPerProof: 12
-            });
-        }
-
         if (_tierId == LibTiers.TIER_SGX) {
             return ITierProvider.Tier({
                 verifierName: "tier_sgx",
-                validityBond: 500 ether, // TKO
-                contestBond: 1000 ether, // TKO
+                validityBond: 250 ether, // TKO
+                contestBond: 1640 ether, // =250TKO * 6.5625
                 cooldownWindow: 1440, //24 hours
                 provingWindow: 60, // 1 hours
                 maxBlocksToVerifyPerProof: 8
+            });
+        }
+
+        if (_tierId == LibTiers.TIER_SGX_ZKVM) {
+            return ITierProvider.Tier({
+                verifierName: "tier_sgx_zkvm",
+                validityBond: 500 ether, // TKO
+                contestBond: 3280 ether, // =500TKO * 6.5625
+                cooldownWindow: 1440, //24 hours
+                provingWindow: 240, // 4 hours
+                maxBlocksToVerifyPerProof: 4
             });
         }
 
@@ -57,16 +57,15 @@ contract TestnetTierProvider is EssentialContract, ITierProvider {
     /// @inheritdoc ITierProvider
     function getTierIds() public pure override returns (uint16[] memory tiers_) {
         tiers_ = new uint16[](3);
-        tiers_[0] = LibTiers.TIER_OPTIMISTIC;
-        tiers_[1] = LibTiers.TIER_SGX;
+        tiers_[0] = LibTiers.TIER_SGX;
+        tiers_[1] = LibTiers.TIER_SGX_ZKVM;
         tiers_[2] = LibTiers.TIER_GUARDIAN;
     }
 
     /// @inheritdoc ITierProvider
     function getMinTier(uint256 _rand) public pure override returns (uint16) {
-        // 10% will be selected to require SGX proofs.
-        if (_rand % 10 == 0) return LibTiers.TIER_SGX;
-        // Other blocks are optimistic, without validity proofs.
-        return LibTiers.TIER_OPTIMISTIC;
+        // 0.1% require SGX + ZKVM; all others require SGX
+        if (_rand % 1000 == 0) return LibTiers.TIER_SGX_ZKVM;
+        else return LibTiers.TIER_SGX;
     }
 }