+++ packages/protocol/contracts/L2/Lib1559Math.sol
@@ -2,6 +2,7 @@
 pragma solidity 0.8.24;
 
 import "../thirdparty/solmate/LibFixedPointMath.sol";
+import "../libs/LibMath.sol";
 
 /// @title Lib1559Math
 /// @notice Implements e^(x) based bonding curve for EIP-1559
@@ -9,10 +10,40 @@ import "../thirdparty/solmate/LibFixedPointMath.sol";
 /// difference as stated in docs/eip1559_on_l2.md.
 /// @custom:security-contact security@taiko.xyz
 library Lib1559Math {
+    using LibMath for uint256;
+
     error EIP1559_INVALID_PARAMS();
 
+    function calc1559BaseFee(
+        uint32 _gasTargetPerL1Block,
+        uint8 _adjustmentQuotient,
+        uint64 _gasExcessMinValue,
+        uint64 _gasExcess,
+        uint64 _gasIssuance,
+        uint32 _parentGasUsed
+    )
+        internal
+        pure
+        returns (uint256 basefee_, uint64 gasExcess_)
+    {
+        // We always add the gas used by parent block to the gas excess
+        // value as this has already happened
+        uint256 excess = uint256(_gasExcess) + _parentGasUsed;
+        excess = excess > _gasIssuance ? excess - _gasIssuance : 1;
+        gasExcess_ = uint64(excess.min(type(uint64).max).max(_gasExcessMinValue));
+
+        // The base fee per gas used by this block is the spot price at the
+        // bonding curve, regardless the actual amount of gas used by this
+        // block, however, this block's gas used will affect the next
+        // block's base fee.
+        basefee_ = basefee(gasExcess_, uint256(_adjustmentQuotient) * _gasTargetPerL1Block);
+
+        // Always make sure basefee is nonzero, this is required by the node.
+        if (basefee_ == 0) basefee_ = 1;
+    }
+
     /// @dev eth_qty(excess_gas_issued) / (TARGET * ADJUSTMENT_QUOTIENT)
-    /// @param _gasExcess TBD
+    /// @param _gasExcess The gas excess value
     /// @param _adjustmentFactor The product of gasTarget and adjustmentQuotient
     function basefee(
         uint256 _gasExcess,
@@ -26,6 +57,8 @@ library Lib1559Math {
             revert EIP1559_INVALID_PARAMS();
         }
 
+        // Note that `/ _adjustmentFactor` can be removed which will make basefee increment faster.
+        // Alternatively, a larger initial value can be used for `_gasExcess`.
         return _ethQty(_gasExcess, _adjustmentFactor) / LibFixedPointMath.SCALING_FACTOR
             / _adjustmentFactor;
     }