+++ packages/protocol/contracts/bridge/Bridge.sol
@@ -4,6 +4,7 @@ pragma solidity 0.8.24;
 import "@openzeppelin/contracts/utils/Address.sol";
 import "../common/EssentialContract.sol";
 import "../libs/LibAddress.sol";
+import "../libs/LibNetwork.sol";
 import "../signal/ISignalService.sol";
 import "./IBridge.sol";
 
@@ -52,7 +53,10 @@ contract Bridge is EssentialContract, IBridge {
     error B_INVALID_STATUS();
     error B_INVALID_USER();
     error B_INVALID_VALUE();
+    error B_MESSAGE_NOT_PROVEN();
     error B_MESSAGE_NOT_SENT();
+    error B_MESSAGE_NOT_SUSPENDED();
+    error B_MESSAGE_SUSPENDED();
     error B_NON_RETRIABLE();
     error B_NOT_FAILED();
     error B_NOT_RECEIVED();
@@ -85,16 +89,32 @@ contract Bridge is EssentialContract, IBridge {
         external
         onlyFromOwnerOrNamed("bridge_watchdog")
     {
-        uint64 _timestamp = _suspend ? type(uint64).max : uint64(block.timestamp);
         for (uint256 i; i < _msgHashes.length; ++i) {
             bytes32 msgHash = _msgHashes[i];
-            proofReceipt[msgHash].receivedAt = _timestamp;
-            emit MessageSuspended(msgHash, _suspend);
+
+            if (_suspend) {
+                if (proofReceipt[msgHash].receivedAt == 0) revert B_MESSAGE_NOT_PROVEN();
+                if (proofReceipt[msgHash].receivedAt == type(uint64).max) {
+                    revert B_MESSAGE_SUSPENDED();
+                }
+
+                proofReceipt[msgHash].receivedAt = type(uint64).max;
+                emit MessageSuspended(msgHash, true, 0);
+            } else {
+                // Note before we set the receivedAt to current timestamp, we have to be really
+                // careful that this message must have been proven then suspended.
+                if (proofReceipt[msgHash].receivedAt != type(uint64).max) {
+                    revert B_MESSAGE_NOT_SUSPENDED();
+                }
+                proofReceipt[msgHash].receivedAt = uint64(block.timestamp);
+                emit MessageSuspended(msgHash, false, uint64(block.timestamp));
+            }
         }
     }
 
     /// @notice Ban or unban an address. A banned addresses will not be invoked upon
     /// with message calls.
+    /// @dev Do not make this function `nonReentrant`, this breaks {DelegateOwner} support.
     /// @param _addr The address to ban or unban.
     /// @param _ban True if ban, false if unban.
     function banAddress(
@@ -103,7 +123,6 @@ contract Bridge is EssentialContract, IBridge {
     )
         external
         onlyFromOwnerOrNamed("bridge_watchdog")
-        nonReentrant
     {
         if (addressBanned[_addr] == _ban) revert B_INVALID_STATUS();
         addressBanned[_addr] = _ban;
@@ -165,9 +184,12 @@ contract Bridge is EssentialContract, IBridge {
         if (messageStatus[msgHash] != Status.NEW) revert B_STATUS_MISMATCH();
 
         uint64 receivedAt = proofReceipt[msgHash].receivedAt;
-        bool isMessageProven = receivedAt != 0;
+        if (receivedAt == type(uint64).max) revert B_MESSAGE_SUSPENDED();
 
-        if (!isMessageProven) {
+        (uint256 invocationDelay,) = getInvocationDelays();
+
+        bool isNewlyProven;
+        if (receivedAt == 0) {
             address signalService = resolve("signal_service", false);
 
             if (!ISignalService(signalService).isSignalSent(address(this), msgHash)) {
@@ -180,10 +202,12 @@ contract Bridge is EssentialContract, IBridge {
             }
 
             receivedAt = uint64(block.timestamp);
-            proofReceipt[msgHash].receivedAt = receivedAt;
-        }
+            isNewlyProven = true;
 
-        (uint256 invocationDelay,) = getInvocationDelays();
+            if (invocationDelay != 0) {
+                proofReceipt[msgHash].receivedAt = receivedAt;
+            }
+        }
 
         if (block.timestamp >= invocationDelay + receivedAt) {
             delete proofReceipt[msgHash];
@@ -205,7 +229,7 @@ contract Bridge is EssentialContract, IBridge {
                 _message.srcOwner.sendEtherAndVerify(_message.value);
             }
             emit MessageRecalled(msgHash);
-        } else if (!isMessageProven) {
+        } else if (isNewlyProven) {
             emit MessageReceived(msgHash, _message, true);
         } else {
             revert B_INVOCATION_TOO_EARLY();
@@ -226,17 +250,20 @@ contract Bridge is EssentialContract, IBridge {
         if (messageStatus[msgHash] != Status.NEW) revert B_STATUS_MISMATCH();
 
         address signalService = resolve("signal_service", false);
+
         uint64 receivedAt = proofReceipt[msgHash].receivedAt;
-        bool isMessageProven = receivedAt != 0;
+        if (receivedAt == type(uint64).max) revert B_MESSAGE_SUSPENDED();
 
         (uint256 invocationDelay, uint256 invocationExtraDelay) = getInvocationDelays();
 
-        if (!isMessageProven) {
+        bool isNewlyProven;
+        if (receivedAt == 0) {
             if (!_proveSignalReceived(signalService, msgHash, _message.srcChainId, _proof)) {
                 revert B_NOT_RECEIVED();
             }
 
             receivedAt = uint64(block.timestamp);
+            isNewlyProven = true;
 
             if (invocationDelay != 0) {
                 proofReceipt[msgHash] = ProofReceipt({
@@ -274,8 +301,8 @@ contract Bridge is EssentialContract, IBridge {
                 refundAmount = _message.value;
                 _updateMessageStatus(msgHash, Status.DONE);
             } else {
-                // Use the specified message gas limit if called by the owner, else
-                // use remaining gas
+                // Use the remaining gas if called by a the destOwner, else
+                // use the specified gas limit.
                 uint256 gasLimit = msg.sender == _message.destOwner ? gasleft() : _message.gasLimit;
 
                 if (_invokeMessageCall(_message, msgHash, gasLimit)) {
@@ -298,7 +325,7 @@ contract Bridge is EssentialContract, IBridge {
                 refundTo.sendEtherAndVerify(refundAmount);
             }
             emit MessageExecuted(msgHash);
-        } else if (!isMessageProven) {
+        } else if (isNewlyProven) {
             emit MessageReceived(msgHash, _message, false);
         } else {
             revert B_INVOCATION_TOO_EARLY();
@@ -417,24 +444,11 @@ contract Bridge is EssentialContract, IBridge {
         virtual
         returns (uint256 invocationDelay_, uint256 invocationExtraDelay_)
     {
-        if (
-            block.chainid == 1 // Ethereum mainnet
-        ) {
-            // For Taiko mainnet
+        if (LibNetwork.isEthereumMainnetOrTestnet(block.chainid)) {
+            // For Taiko mainnet and public testnets
             // 384 seconds = 6.4 minutes = one ethereum epoch
             return (1 hours, 384 seconds);
-        } else if (
-            block.chainid == 2 // Ropsten
-                || block.chainid == 4 // Rinkeby
-                || block.chainid == 5 // Goerli
-                || block.chainid == 42 // Kovan
-                || block.chainid == 17_000 // Holesky
-                || block.chainid == 11_155_111 // Sepolia
-        ) {
-            // For all Taiko public testnets
-            return (30 minutes, 384 seconds);
-        } else if (block.chainid >= 32_300 && block.chainid <= 32_400) {
-            // For all Taiko internal devnets
+        } else if (LibNetwork.isTaikoDevnet(block.chainid)) {
             return (5 minutes, 384 seconds);
         } else {
             // This is a Taiko L2 chain where no deleys are applied.
@@ -454,14 +468,18 @@ contract Bridge is EssentialContract, IBridge {
         return _msgHash ^ bytes32(uint256(Status.FAILED));
     }
 
-    /// @notice Checks if the given address can pause and unpause the bridge.
-    function _authorizePause(address)
-        internal
-        view
-        virtual
-        override
-        onlyFromOwnerOrNamed("bridge_pauser")
-    { }
+    /// @notice Checks if the given address can pause and/or unpause the bridge.
+    /// @dev Considering that the watchdog is a hot wallet, in case its private key is leaked, we
+    /// only allow watchdog to pause the bridge, but does not allow it to unpause the bridge.
+    function _authorizePause(address addr, bool toPause) internal view virtual override {
+        // Owenr and chain_pauser can pause/unpause the bridge.
+        if (addr == owner() || addr == resolve("chain_pauser", true)) return;
+
+        // bridge_watchdog can pause the bridge, but cannot unpause it.
+        if (toPause && addr == resolve("bridge_watchdog", true)) return;
+
+        revert RESOLVER_DENIED();
+    }
 
     /// @notice Invokes a call message on the Bridge.
     /// @param _message The call message to be invoked.
@@ -518,7 +536,7 @@ contract Bridge is EssentialContract, IBridge {
 
     /// @notice Resets the call context
     function _resetContext() private {
-        if (block.chainid == 1) {
+        if (LibNetwork.isDencunSupported(block.chainid)) {
             _storeContext(bytes32(0), address(0), uint64(0));
         } else {
             _storeContext(bytes32(PLACEHOLDER), address(uint160(PLACEHOLDER)), uint64(PLACEHOLDER));
@@ -530,7 +548,7 @@ contract Bridge is EssentialContract, IBridge {
     /// @param _from The sender's address.
     /// @param _srcChainId The source chain ID.
     function _storeContext(bytes32 _msgHash, address _from, uint64 _srcChainId) private {
-        if (block.chainid == 1) {
+        if (LibNetwork.isDencunSupported(block.chainid)) {
             assembly {
                 tstore(_CTX_SLOT, _msgHash)
                 tstore(add(_CTX_SLOT, 1), _from)
@@ -544,7 +562,7 @@ contract Bridge is EssentialContract, IBridge {
     /// @notice Loads and returns the call context.
     /// @return ctx_ The call context.
     function _loadContext() private view returns (Context memory) {
-        if (block.chainid == 1) {
+        if (LibNetwork.isDencunSupported(block.chainid)) {
             bytes32 msgHash;
             address from;
             uint64 srcChainId;
@@ -574,10 +592,12 @@ contract Bridge is EssentialContract, IBridge {
         private
         returns (bool)
     {
-        bytes memory data = abi.encodeCall(
-            ISignalService.proveSignalReceived,
-            (_chainId, resolve(_chainId, "bridge", false), _signal, _proof)
-        );
-        return _signalService.sendEther(0, gasleft(), data);
+        try ISignalService(_signalService).proveSignalReceived(
+            _chainId, resolve(_chainId, "bridge", false), _signal, _proof
+        ) {
+            return true;
+        } catch {
+            return false;
+        }
     }
 }